Move the gunicorn setup out to a separate file
