+++ routemaster/cli.py
@@ -1,11 +1,11 @@
 """CLI handling for `routemaster`."""
 import yaml
 import click
-import gunicorn.app.base
 
 from routemaster.app import App
 from routemaster.config import load_config
 from routemaster.server import server
+from routemaster.gunicorn_application import HackyWSGIApplication
 
 
 @click.group()
@@ -48,24 +48,6 @@ def validate(ctx):
 def serve(ctx, bind, debug):
     """Entrypoint for serving the Routemaster HTTP service."""
     server.config.app = ctx.obj
-
-    class HackyWSGIApplication(gunicorn.app.base.BaseApplication):
-        def __init__(self, app, *, bind, debug):
-            self.application = app
-            self.bind = bind
-            self.debug = debug
-            super().__init__()
-
-        def load_config(self):
-            self.cfg.set('bind', self.bind)
-            self.cfg.set('workers', 1)
-
-            if self.debug:
-                self.cfg.set('reload', True)
-                self.cfg.set('accesslog', '-')
-
-        def load(self):
-            return self.application
 
     instance = HackyWSGIApplication(server, bind=bind, debug=debug)
     instance.run()
